refactor: 토큰 생성 시 memberId 활용
