+++ src/main/java/roomescape/admin/dto/ReservationFilterRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.admin.dto;
 
 import java.time.LocalDate;
