+++ src/main/java/roomescape/member/dto/LoginRequest.java
@@ -3,27 +3,13 @@ package roomescape.member.dto;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
-public class LoginRequest {
-    @NotBlank
-    @Email
-    private String email;
+public record LoginRequest (
+        @NotBlank
+        @Email
+        String email,
 
-    @NotBlank
-    private String password;
+        @NotBlank
+        String password
 
-    public LoginRequest() {
-    }
-
-    public LoginRequest(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
+){
 }
