+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -3,7 +3,7 @@ package roomescape.reservation.service;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.dto.ReservationFilterRequest;
+import roomescape.admin.dto.ReservationFilterRequest;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.repository.MemberRepository;
 import roomescape.member.dto.LoginMember;
